LGEQualcommRIL: Fix PIN status for e73x
The QUERY_FACILITY_LOCK call is completely broken on the e730/739
when the SIM card status (SC) is requested.
Instead of letting it go through (and fail), use the pin1 status
from the IccCard application to figure out what the lock state is,
and return it as a message to the calling handler.

Change-Id: Iaed16bcb1a3ee71d9e833e4a4ccee423de0593a4