SearchWidget: passing APP_DATA to voice intent (from ics)
ics: http://review.cyanogenmod.com/#/c/17826/1

SearchView uses SearchManager.APP_DATA to pass data between the origin activity
and the search activity. This works for text searches (keyboard), but it's not
implemented by voice recognition searches. The method createVoiceAppSearchIntent
initializes a empty bundle that is filled with the QUERY parameter.
This patch adds the SearchManager.APP_DATA parameter if is present, let the
search activity access to the data passed by the origin activity

Change-Id: I2d0d853b69b36aa4e52f3cd49b4364f5fbb5ab93